finetune-qwen-master/requirements.txt
absl-py==2.1.0
accelerate==1.5.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrdict==2.0.1
attrs==25.3.0
av==14.2.0
binpacking==1.5.2
black==24.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dacite==1.9.2
datasets==3.2.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1741148399929/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
decord==0.6.0
deepspeed==0.16.5
dill==0.3.8
distro==1.9.0
einops==0.8.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flash_attn==2.7.4.post1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
iniconfig==2.1.0
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1741457126/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.12.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.8
lazy_loader==0.4
libcst==1.7.0
loguru==0.7.3
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
modelscope==1.23.2
moreorless==0.4.0
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1741006472358/work
mpmath==1.3.0
ms_swift==3.4.1.post1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.66.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
oss2==2.19.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathspec==0.12.1
peft==0.14.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pluggy==1.6.0
portalocker==3.1.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.3.0
protobuf==6.30.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663123172/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing==3.2.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1741805125248/work
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=sam2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
sortedcontainers==2.4.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.1
stdlibs==2025.4.4
stringzilla==3.12.3
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615905931/work
tqdm==4.67.1
trailrunner==1.4.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.1
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.15.2
typer==0.15.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
ufmt==2.0.0b2
urllib3==2.3.0
usort==1.0.2
uvicorn==0.34.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
